Refactored in-line editing of lists
